requirements.txt
Django==4.1.7
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-filter==22.1
python-decouple